step-002: extract shared desktop contract types
