+++ docs/plans/2026-03-15-openwork-desktop-tauri-to-electron-migration/steps.json
@@ -30,7 +30,7 @@
         "The extracted module is Electron-neutral in naming and location.",
         "No type shape changes are introduced in this extraction step."
       ],
-      "completed": false
+      "completed": true
     },
     {
       "serialNumber": "003",
