+++ packages/app/src/app/lib/desktop-contract.ts
@@ -1,0 +1,352 @@
+export type EngineInfo = {
+  running: boolean;
+  runtime: "direct" | "openwork-orchestrator";
+  baseUrl: string | null;
+  projectDir: string | null;
+  hostname: string | null;
+  port: number | null;
+  opencodeUsername: string | null;
+  opencodePassword: string | null;
+  pid: number | null;
+  lastStdout: string | null;
+  lastStderr: string | null;
+};
+
+export type OpenworkServerInfo = {
+  running: boolean;
+  host: string | null;
+  port: number | null;
+  baseUrl: string | null;
+  connectUrl: string | null;
+  mdnsUrl: string | null;
+  lanUrl: string | null;
+  clientToken: string | null;
+  hostToken: string | null;
+  pid: number | null;
+  lastStdout: string | null;
+  lastStderr: string | null;
+};
+
+export type OrchestratorDaemonState = {
+  pid: number;
+  port: number;
+  baseUrl: string;
+  startedAt: number;
+};
+
+export type OrchestratorOpencodeState = {
+  pid: number;
+  port: number;
+  baseUrl: string;
+  startedAt: number;
+};
+
+export type OrchestratorBinaryInfo = {
+  path: string;
+  source: string;
+  expectedVersion?: string | null;
+  actualVersion?: string | null;
+};
+
+export type OrchestratorBinaryState = {
+  opencode?: OrchestratorBinaryInfo | null;
+};
+
+export type OrchestratorSidecarInfo = {
+  dir?: string | null;
+  baseUrl?: string | null;
+  manifestUrl?: string | null;
+  target?: string | null;
+  source?: string | null;
+  opencodeSource?: string | null;
+  allowExternal?: boolean | null;
+};
+
+export type OrchestratorWorkspace = {
+  id: string;
+  name: string;
+  path: string;
+  workspaceType: string;
+  baseUrl?: string | null;
+  directory?: string | null;
+  createdAt?: number | null;
+  lastUsedAt?: number | null;
+};
+
+export type OrchestratorStatus = {
+  running: boolean;
+  dataDir: string;
+  daemon: OrchestratorDaemonState | null;
+  opencode: OrchestratorOpencodeState | null;
+  cliVersion?: string | null;
+  sidecar?: OrchestratorSidecarInfo | null;
+  binaries?: OrchestratorBinaryState | null;
+  activeId: string | null;
+  workspaceCount: number;
+  workspaces: OrchestratorWorkspace[];
+  lastError: string | null;
+};
+
+export type EngineDoctorResult = {
+  found: boolean;
+  inPath: boolean;
+  resolvedPath: string | null;
+  version: string | null;
+  supportsServe: boolean;
+  notes: string[];
+  serveHelpStatus: number | null;
+  serveHelpStdout: string | null;
+  serveHelpStderr: string | null;
+};
+
+export type WorkspaceInfo = {
+  id: string;
+  name: string;
+  path: string;
+  preset: string;
+  workspaceType: "local" | "remote";
+  remoteType?: "openwork" | "opencode" | null;
+  baseUrl?: string | null;
+  directory?: string | null;
+  displayName?: string | null;
+  openworkHostUrl?: string | null;
+  openworkToken?: string | null;
+  openworkWorkspaceId?: string | null;
+  openworkWorkspaceName?: string | null;
+
+  // Sandbox lifecycle metadata (desktop-managed)
+  sandboxBackend?: "docker" | null;
+  sandboxRunId?: string | null;
+  sandboxContainerName?: string | null;
+};
+
+export type WorkspaceList = {
+  activeId: string;
+  workspaces: WorkspaceInfo[];
+};
+
+export type WorkspaceExportSummary = {
+  outputPath: string;
+  included: number;
+  excluded: string[];
+};
+
+export type OpencodeCommandDraft = {
+  name: string;
+  description?: string;
+  template: string;
+  agent?: string;
+  model?: string;
+  subtask?: boolean;
+};
+
+export type WorkspaceOpenworkConfig = {
+  version: number;
+  workspace?: {
+    name?: string | null;
+    createdAt?: number | null;
+    preset?: string | null;
+  } | null;
+  authorizedRoots: string[];
+  reload?: {
+    auto?: boolean;
+    resume?: boolean;
+  } | null;
+};
+
+export type AppBuildInfo = {
+  version: string;
+  gitSha?: string | null;
+  buildEpoch?: string | null;
+  openworkDevMode?: boolean;
+};
+
+export type OrchestratorDetachedHost = {
+  openworkUrl: string;
+  token: string;
+  hostToken: string;
+  port: number;
+  sandboxBackend?: "docker" | null;
+  sandboxRunId?: string | null;
+  sandboxContainerName?: string | null;
+};
+
+export type SandboxDoctorResult = {
+  installed: boolean;
+  daemonRunning: boolean;
+  permissionOk: boolean;
+  ready: boolean;
+  clientVersion?: string | null;
+  serverVersion?: string | null;
+  error?: string | null;
+  debug?: {
+    candidates: string[];
+    selectedBin?: string | null;
+    versionCommand?: {
+      status: number;
+      stdout: string;
+      stderr: string;
+    } | null;
+    infoCommand?: {
+      status: number;
+      stdout: string;
+      stderr: string;
+    } | null;
+  } | null;
+};
+
+export type OpenworkDockerCleanupResult = {
+  candidates: string[];
+  removed: string[];
+  errors: string[];
+};
+
+export type SandboxDebugProbeResult = {
+  startedAt: number;
+  finishedAt: number;
+  runId: string;
+  workspacePath: string;
+  ready: boolean;
+  doctor: SandboxDoctorResult;
+  detachedHost?: OrchestratorDetachedHost | null;
+  dockerInspect?: {
+    status: number;
+    stdout: string;
+    stderr: string;
+  } | null;
+  dockerLogs?: {
+    status: number;
+    stdout: string;
+    stderr: string;
+  } | null;
+  cleanup: {
+    containerName?: string | null;
+    containerRemoved: boolean;
+    removeResult?: {
+      status: number;
+      stdout: string;
+      stderr: string;
+    } | null;
+    workspaceRemoved: boolean;
+    errors: string[];
+  };
+  error?: string | null;
+};
+
+export type ExecResult = {
+  ok: boolean;
+  status: number;
+  stdout: string;
+  stderr: string;
+};
+
+export type ScheduledJobRun = {
+  prompt?: string;
+  command?: string;
+  arguments?: string;
+  files?: string[];
+  agent?: string;
+  model?: string;
+  variant?: string;
+  title?: string;
+  share?: boolean;
+  continue?: boolean;
+  session?: string;
+  runFormat?: string;
+  attachUrl?: string;
+  port?: number;
+};
+
+export type ScheduledJob = {
+  scopeId?: string;
+  timeoutSeconds?: number;
+  invocation?: { command: string; args: string[] };
+  slug: string;
+  name: string;
+  schedule: string;
+  prompt?: string;
+  attachUrl?: string;
+  run?: ScheduledJobRun;
+  source?: string;
+  workdir?: string;
+  createdAt: string;
+  updatedAt?: string;
+  lastRunAt?: string;
+  lastRunExitCode?: number;
+  lastRunError?: string;
+  lastRunSource?: string;
+  lastRunStatus?: string;
+};
+
+export type LocalSkillCard = {
+  name: string;
+  path: string;
+  description?: string;
+  trigger?: string;
+};
+
+export type LocalSkillContent = {
+  path: string;
+  content: string;
+};
+
+export type OpencodeConfigFile = {
+  path: string;
+  exists: boolean;
+  content: string | null;
+};
+
+export type UpdaterEnvironment = {
+  supported: boolean;
+  reason: string | null;
+  executablePath: string | null;
+  appBundlePath: string | null;
+};
+
+export type CacheResetResult = {
+  removed: string[];
+  missing: string[];
+  errors: string[];
+};
+
+export type ObsidianMirrorFileContent = {
+  exists: boolean;
+  path: string;
+  content: string | null;
+  updatedAtMs: number | null;
+};
+
+// OpenCodeRouter types
+export type OpenCodeRouterIdentityItem = {
+  id: string;
+  enabled: boolean;
+  running?: boolean;
+};
+
+export type OpenCodeRouterChannelStatus = {
+  items: OpenCodeRouterIdentityItem[];
+};
+
+export type OpenCodeRouterStatus = {
+  running: boolean;
+  config: string;
+  healthPort?: number | null;
+  telegram: OpenCodeRouterChannelStatus;
+  slack: OpenCodeRouterChannelStatus;
+  opencode: { url: string; directory?: string };
+};
+
+export type OpenCodeRouterStatusResult =
+  | { ok: true; status: OpenCodeRouterStatus }
+  | { ok: false; error: string };
+
+export type OpenCodeRouterInfo = {
+  running: boolean;
+  version: string | null;
+  workspacePath: string | null;
+  opencodeUrl: string | null;
+  healthPort: number | null;
+  pid: number | null;
+  lastStdout: string | null;
+  lastStderr: string | null;
+};
